requirements.txt
# Core dependencies
aiohttp>=3.9.1
python-telegram-bot>=20.7
python-dotenv>=1.0.0
firebase-admin>=6.2.0
beautifulsoup4>=4.12.2
feedparser>=6.0.11
requests>=2.31.0
httpx[http2]>=0.24.1  # For RemoteOK and Remotive fetchers
google-cloud-firestore>=2.12.0  # For Firestore async support

# Development dependencies
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.1
python-dateutil>=2.8.2
